using double to account for decimals from division

+++ calc.c
@@ -2,9 +2,9 @@
 #include <stdlib.h>
 #include <string.h>
 
-int parseFactor(char **expression) {
+double parseFactor(char **expression) {
   if (**expression >= '0' && **expression <= '9') {
-    int result = **expression - '0';
+    double result = **expression - '0';
     (*expression)++;
     return result;
   } else if (**expression == '-') {
@@ -16,36 +16,34 @@ int parseFactor(char **expression) {
   return 0;
 }
 
-int parseProduct(char **expression) {
-  int fac1 = parseFactor(expression);
+double parseProduct(char **expression) {
+  double fac1 = parseFactor(expression);
   while (**expression == '*' || **expression == '/') {
     int mulOrDiv = 1;
     if (**expression == '/')
       mulOrDiv = 0;
 
     ++(*expression);
-    int fac2 = parseFactor(expression);
+    double fac2 = parseFactor(expression);
     if (mulOrDiv == 1) {
       fac1 = fac1 * fac2;
     } else {
-
       fac1 = fac1 / fac2;
     }
   }
   return fac1;
 };
 
-int parseSum(char **expression) {
-  int fac1 = parseProduct(expression);
+double parseSum(char **expression) {
+  double fac1 = parseProduct(expression);
 
   while (**expression == '+' || **expression == '-') {
-    // Addition is default 1; Subtraction is 0;
     int addOrSub = 1;
     if (**expression == '-')
       addOrSub = 0;
 
     ++(*expression);
-    int fac2 = parseProduct(expression);
+    double fac2 = parseProduct(expression);
     if (addOrSub == 1) {
       fac1 = fac1 + fac2;
     } else {
@@ -59,8 +57,8 @@ int parseSum(char **expression) {
 int main(int argc, char *argv[]) {
   char *expression = argv[1];
 
-  int result = parseSum(&expression);
+  double result = parseSum(&expression);
 
-  printf("Result: %d\n", result);
+  printf("Result: %.2f\n", result);
   return 0;
 }
